Code Structure Changes
The tight program architecture means that certain errors disrupt the
program severely. I am going to begin fixing that in the next few days.
I am hoping to get a basic alpha running before the end of the month

Also I added memory buttons. The form is now full but will be expanded
in the future